+++ CHANGES.md
@@ -19,5 +19,6 @@ Apollo 2.5.0
 * [Security: Hide password when registering or modifying users](https://github.com/apolloconfig/apollo/pull/5414)
 * [Fix: the logical judgment for configuration addition, deletion, and modification.](https://github.com/apolloconfig/apollo/pull/5432)
 * [Feature support incremental configuration synchronization client](https://github.com/apolloconfig/apollo/pull/5288)
+* [Fix: Replace synchronized Multimap with case-insensitive concurrent implementation to eliminate thread blocking in NotificationControllerV2](https://github.com/apolloconfig/apollo/pull/5450)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/16?closed=1)